feat: expose serialisation functions
In developing bb-rs, we want to be able to serialise bytecode directly
to bincode (and possibly msgpack) rather than roundtripping through gzip
and/or base64. Therefore it would be very helpful to be able to use
bincode_serialize et al directly.

This commit simply adds those methods to the public API.